create workflow project from REST API instead of GQL
