+++ app.py
@@ -1,6 +1,6 @@
 import logging
 import graphene
-from api.project import ProjectDetails
+from api.project import ProjectDetails, ProjectCreationFromWorkflow
 from api.transfer import (
     AssociationsImport,
     FileExport,
@@ -40,6 +40,7 @@ routes = [
     Route("/project/{project_id:str}/import_file", PrepareFileImport),
     Route("/project/{project_id:str}/import_json", JSONImport),
     Route("/project/{project_id:str}/import/task/{task_id:str}", UploadTask),
+    Route("/project", ProjectCreationFromWorkflow),  # create fro workflow
 ]
 
 middleware = [Middleware(DatabaseSessionHandler)]
