+++ graphql_api/mutation/project.py
@@ -1,36 +1,14 @@
-import os
 from typing import Optional
-from controller.tokenization import tokenization_service
 
 import graphene
 
 from controller.auth import manager as auth
-from graphql_api import types
 from graphql_api.types import Project
 from submodules.model import enums, events
 from controller.project import manager
-from submodules.s3.controller import bucket_exists, create_bucket
 from util import doc_ock, notification
 from submodules.model.business_objects import notification as notification_model
-
-
-from controller.transfer.record_transfer_manager import import_records_and_rlas
-from controller.transfer.manager import check_and_add_running_id
-from controller.auth import manager as auth_manager
-from controller.project import manager as project_manager
-from controller.upload_task import manager as upload_task_manager
 from submodules.model import enums
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-
-__engine = None
-
-
-def get_workflow_db_engine():
-    global __engine
-    if __engine is None:
-        __engine = create_engine(os.getenv("WORKFLOW_POSTGRES"))
-    return __engine
 
 
 class CreateProject(graphene.Mutation):
@@ -56,67 +34,6 @@ class CreateProject(graphene.Mutation):
             events.CreateProject(Name=f"{name}-{project.id}", Description=description),
         )
         return CreateProject(project=project, ok=True)
-
-
-class CreateProjectByWorkflowStore(graphene.Mutation):
-    class Arguments:
-        name = graphene.String(required=True)
-        description = graphene.String(required=True)
-        tokenizer = graphene.String(required=True)
-        storeId = graphene.String(required=True)
-
-    ok = graphene.Boolean()
-    project = graphene.Field(lambda: Project)
-
-    def mutate(self, info, name: str, description: str, tokenizer: str, storeId: str):
-        user = auth.get_user_by_info(info)
-        organization = auth_manager.get_organization_by_user_id(user.id)
-
-        if not bucket_exists(str(organization.id)):
-            create_bucket(str(organization.id))
-
-        project = project_manager.create_project(
-            organization.id, name, description, user.id
-        )
-        project_manager.update_project(project_id=project.id, tokenizer=tokenizer)
-
-        Session = sessionmaker(get_workflow_db_engine())
-        with Session() as session:
-            results = session.execute(
-                f"SELECT record FROM store_entry WHERE store_id = '{storeId}'"
-            ).all()
-
-        data = [result for result, in results]
-
-        upload_task = upload_task_manager.create_upload_task(
-            user_id=user.id,
-            project_id=project.id,
-            file_name=name,
-            file_type="json",
-            file_import_options="",
-            upload_type=enums.UploadTypes.WORKFLOW_STORE.value,
-        )
-        import_records_and_rlas(
-            project.id,
-            user.id,
-            data,
-            upload_task,
-            enums.RecordCategory.SCALE.value,
-        )
-        check_and_add_running_id(project.id, user.id)
-
-        upload_task_manager.update_upload_task_to_finished(upload_task)
-        tokenization_service.request_tokenize_project(str(project.id), str(user.id))
-
-        notification.send_organization_update(
-            project.id, f"project_created:{str(project.id)}", True
-        )
-        doc_ock.post_event(
-            user,
-            events.CreateProject(Name=f"{name}-{project.id}", Description=description),
-        )
-
-        return CreateProjectByWorkflowStore(project=project, ok=True)
 
 
 class CreateSampleProject(graphene.Mutation):
@@ -224,7 +141,6 @@ class UpdateProjectTokenizer(graphene.Mutation):
 
 class ProjectMutation(graphene.ObjectType):
     create_project = CreateProject.Field()
-    create_project_by_workflow_store = CreateProjectByWorkflowStore.Field()
     create_sample_project = CreateSampleProject.Field()
     delete_project = DeleteProject.Field()
     update_project_status = UpdateProjectStatus.Field()
